REGRESSION(r274244): [GTK][WPE] Two http/tests/security/contentSecurityPolicy tests crash https://bugs.webkit.org/show_bug.cgi?id=223978
Reviewed by Philippe Normand.

Apache not longer sends a Content-Length header with the cgi/python version of this tests,
but it used chunked transfer encoding.
It seems this confuses libsoup causing a critical event.
Add an empty \r\n at the end of the test to indicate that the server has finished sending
the data.

* http/tests/security/contentSecurityPolicy/1.1/scripthash-allowed-by-legacy-enforced-policy-and-blocked-by-report-policy.py:
* http/tests/security/contentSecurityPolicy/1.1/scripthash-blocked-by-legacy-enforced-policy-and-blocked-by-report-policy.py:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275274 268f45cc-cd09-0410-ab3c-d52691b4dbfc