Animated pseudo element style resolved against wrong parent style https://bugs.webkit.org/show_bug.cgi?id=223990 rdar://74997361
Reviewed by Antoine Quint.
Source/WebCore:

In createAnimatedElementUpdate we get the parent and parent box styles from the parent stack.
This is wrong for pseudo elements. Their parent style should the host style which is not pushed to the stack.

This matters in style adjuster which may apply wrong adjustments as a result.

Test: fast/animation/pseudo-element-style-adjuster.html

* style/StyleTreeResolver.cpp:
(WebCore::Style::TreeResolver::resolveElement):
(WebCore::Style::TreeResolver::resolvePseudoStyle):
(WebCore::Style::TreeResolver::createAnimatedElementUpdate):

Make static and provide the parent and parent box styles as parameters.

* style/StyleTreeResolver.h:

LayoutTests:

* fast/animation/pseudo-element-style-adjuster-expected.html: Added.
* fast/animation/pseudo-element-style-adjuster.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275277 268f45cc-cd09-0410-ab3c-d52691b4dbfc