[iOS] '-webkit-appearance: none' is not respected for searchfield decorations https://bugs.webkit.org/show_bug.cgi?id=224018
Reviewed by Antti Koivisto.

Source/WebCore:

Specifying '-webkit-appearance: none' on a searchfield decoration
pseudo-element fails to hide the native icon. This issue arises from
the fact that we add the icon in the UA stylesheet, rather than painting
it in the theme.

To fix, remove the styles from the UA stylesheet, and paint the icon in
RenderThemeIOS. If a different '-webkit-appearance' is specified the
theme will not adjust styles or paint the icon.

Test: fast/forms/ios/form-control-refresh/search/search-decoration-appearance.html

* css/html.css:

Remove the UA styles for searchfield decorations so that they are not
applied when specifying a different -webkit-appearance.

* rendering/RenderThemeIOS.h:
* rendering/RenderThemeIOS.mm:
(WebCore::RenderThemeIOS::adjustSearchFieldDecorationPartStyle const):

Adjust the style to match the native appearance.

(WebCore::RenderThemeIOS::paintSearchFieldDecorationPart):

Paint the native icon.

(WebCore::RenderThemeIOS::adjustSearchFieldResultsDecorationPartStyle const):
(WebCore::RenderThemeIOS::paintSearchFieldResultsDecorationPart):
(WebCore::RenderThemeIOS::adjustSearchFieldResultsButtonStyle const):
(WebCore::RenderThemeIOS::paintSearchFieldResultsButton):

LayoutTests:

Rebaselined tests to account for the change to the way we paint
searchfield decorations on iOS. Added a new test to verify that
'-webkit-appearance: none' is respected.

* fast/forms/ios/form-control-refresh/search/search-decoration-appearance-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/search/search-decoration-appearance.html: Added.
* platform/ios-simulator/fast/forms/datalist/datalist-searchinput-appearance-expected.txt:
* platform/ios/fast/css/focus-ring-exists-for-search-field-expected.txt:
* platform/ios/fast/css/input-search-padding-expected.txt:
* platform/ios/fast/css/text-input-with-webkit-border-radius-expected.txt:
* platform/ios/fast/css/text-overflow-input-expected.txt:
* platform/ios/fast/forms/box-shadow-override-expected.txt:
* platform/ios/fast/forms/control-restrict-line-height-expected.txt:
* platform/ios/fast/forms/input-appearance-height-expected.txt:
* platform/ios/fast/forms/placeholder-pseudo-style-expected.txt:
* platform/ios/fast/forms/search-cancel-button-style-sharing-expected.txt:
* platform/ios/fast/forms/search-display-none-cancel-button-expected.txt:
* platform/ios/fast/forms/search-input-rtl-expected.txt:
* platform/ios/fast/forms/search-styled-expected.txt:
* platform/ios/fast/replaced/width100percent-searchfield-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275355 268f45cc-cd09-0410-ab3c-d52691b4dbfc