Add a missing bounds check when mapping display list items from a shared display list handle https://bugs.webkit.org/show_bug.cgi?id=224156 <rdar://problem/71805209>
Reviewed by Tim Horton.

Add a check to ensure that we don't end up with out-of-bounds memory access when attempting to map an `offset`
and `capacity` in shared display list data into a `WebCore::DisplayList`. It's possible for this to happen if
either the sum of `offset` and `capacity` (i.e. the extent of the mapped segment) overflows `size_t`, or the
extent exceeds the total length of the buffer in shared memory.

* GPUProcess/graphics/DisplayListReaderHandle.cpp:
(WebKit::DisplayListReaderHandle::displayListForReading const):

Note that returning `nullptr` here subsequently causes us to terminate the web process due to the message check
for `displayList` inside `nextDestinationImageBufferAfterApplyingDisplayLists`.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275473 268f45cc-cd09-0410-ab3c-d52691b4dbfc