[GTK][WPE] Wrong frame scrolled when view is horizontally scrolled with async scrolling enabled https://bugs.webkit.org/show_bug.cgi?id=222900
Patch by Alejandro G. Castro <alex@igalia.com> on 2021-04-07
Reviewed by Žan Doberšek.

Source/WebCore:

We need to use the scrolling offset in the parent when collecting
the descendant layers at a point. We substract the boundsOrigin to
make sure we transform the point considering the scrolling offset.

Test: fast/scrolling/scrolling-inside-scrolled-overflowarea.html

* page/scrolling/nicosia/ScrollingTreeNicosia.cpp:
(WebCore::collectDescendantLayersAtPoint):

Tools:

We were not testing the async code path for WPE, it is important
we do it because it is the default option.

* WebKitTestRunner/wpe/TestControllerWPE.cpp:
(WTR::TestController::platformSpecificFeatureDefaultsForTest const):
Activating AsyncOverflowScrollingEnabled.

LayoutTests:

Add test for the use case.

* fast/scrolling/scrolling-inside-scrolled-overflowarea-expected.txt: Added.
* fast/scrolling/scrolling-inside-scrolled-overflowarea.html: Added.
* platform/ios/TestExpectations: The test uses mouseMoveTo and it is not supported in IOS.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275593 268f45cc-cd09-0410-ab3c-d52691b4dbfc